+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fixed: Bulk request conflict by adding a retry count
 * Added: Edge.getVertices which accepts fetch hints
 * Changed: avoid saving empty mutations in Graph#saveElementMutations 
+* Added: support to read multiple StreamingPropertyValues with on query
 
 # v2.5.1